move instantiation of objects into index file, removing confusing double definitions and usages
