+++ src/commands/utility/create-text-channel.ts
@@ -10,7 +10,7 @@ import {
   SlashCommandStringOption,
   TextChannel,
 } from "discord.js";
-import signups from "../../services/signups";
+import { signups } from "../../services";
 
 module.exports = {
   data: new SlashCommandBuilder()
